@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "serve": "ng serve",
+    "serve": "ng serve --aot",
     "build": "ng build",
     "build-prod": "ng build --prod --build-optimizer && npm run generate-sw",
     "test": "ng test",
@@ -15,32 +15,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0-rc.1",
+    "@angular/animations": "^5.0.0",
     "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^5.0.0-rc.1",
-    "@angular/compiler": "^5.0.0-rc.1",
-    "@angular/core": "^5.0.0-rc.1",
-    "@angular/forms": "^5.0.0-rc.1",
-    "@angular/http": "^5.0.0-rc.1",
-    "@angular/platform-browser": "^5.0.0-rc.1",
-    "@angular/platform-browser-dynamic": "^5.0.0-rc.1",
-    "@angular/router": "^5.0.0-rc.1",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "autotrack": "^2.4.1",
     "core-js": "^2.5.1",
-    "material-components-web": "^0.22.0",
+    "material-components-web": "^0.24.0",
     "ng-lazyload-image": "^3.3.3",
-    "rxjs": "^5.4.3",
+    "rxjs": "^5.5.2",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "v1.5.0-beta.4",
-    "@angular/compiler-cli": "^5.0.0-rc.1",
-    "@angular/language-service": "^5.0.0-rc.1",
-    "@types/jasmine": "^2.5.54",
+    "@angular/cli": "^1.5.0",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/language-service": "^5.0.0",
+    "@types/jasmine": "^2.6.2",
     "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^8.0.34",
-    "codelyzer": "^3.2.1",
-    "firebase-tools": "^3.13.1",
+    "@types/node": "^8.0.47",
+    "codelyzer": "^4.0.0",
+    "firebase-tools": "^3.14.0",
     "jasmine-core": "^2.8.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "^1.7.1",
@@ -50,10 +50,10 @@
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "node-sass": "^4.5.3",
-    "protractor": "~5.1.2",
+    "protractor": "^5.2.0",
     "ts-node": "^3.3.0",
-    "tslint": "^5.7.0",
-    "typescript": "^2.5.3",
+    "tslint": "^5.8.0",
+    "typescript": "^2.6.0",
     "workbox-build": "^2.1.0",
     "workbox-sw": "^2.1.0"
   }
